requirements.txt
# Core scientific stack
numpy>=2.0.0,<2.2
pandas>=2.2.0,<2.3
pyarrow>=15.0.0

# Backtesting & factor libraries
vectorbt>=0.26.0
git+https://github.com/xgboosted/pandas-ta-classic.git@main

# Speed-ups
numba>=0.60
llvmlite>=0.43
bottleneck>=1.4

# Plotting & reporting
matplotlib>=3.8
jinja2>=3.1

# File and config
pyyaml>=6.0

# Optional (for Excel / reports)
openpyxl>=3.1
xlsxwriter>=3.2
duckdb==1.4.0
tabulate>=0.9
pandas-ta-classic>=0.3.14b1
